About this project
#Description
This is a chat app. It is created with firebase , react and some react dependences. First you need to create an account, then you can log-in in the account that you created. When you’re currently logged in you can see your profile, then when you go to home page, and send a message or photo to other users. It is private and fully responsible.
##Deployed project link  https://peaceful-fermi-715236.netlify.app